+++ Meta/Lagom/Fuzzers/FuzzTTF.cpp
@@ -4,12 +4,12 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
-#include <LibGfx/Font/TrueType/Font.h>
+#include <LibGfx/Font/OpenType/Font.h>
 #include <stddef.h>
 #include <stdint.h>
 
 extern "C" int LLVMFuzzerTestOneInput(u8 const* data, size_t size)
 {
-    (void)TTF::Font::try_load_from_externally_owned_memory({ data, size });
+    (void)OpenType::Font::try_load_from_externally_owned_memory({ data, size });
     return 0;
 }
