+++ Tests/LibTTF/TestCmap.cpp
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
-#include <LibGfx/Font/TrueType/Cmap.h>
+#include <LibGfx/Font/OpenType/Cmap.h>
 #include <LibTest/TestCase.h>
 
 TEST_CASE(test_cmap_format_4)
@@ -54,7 +54,7 @@ TEST_CASE(test_cmap_format_4)
         0, 0,
     };
     // clang-format on
-    auto cmap = TTF::Cmap::from_slice({ cmap_table, sizeof cmap_table }).value();
+    auto cmap = OpenType::Cmap::from_slice({ cmap_table, sizeof cmap_table }).value();
     cmap.set_active_index(0);
 
     // Format 4 can't handle code points > 0xffff.
@@ -74,7 +74,7 @@ TEST_CASE(test_cmap_format_4)
     // From https://docs.microsoft.com/en-us/typography/opentype/spec/cmap#format-4-segment-mapping-to-delta-values:
     // "the final start code and endCode values must be 0xFFFF. This segment need not contain any valid mappings.
     // (It can just map the single character code 0xFFFF to missingGlyph). However, the segment must be present."
-    // FIXME: Make TTF::Cmap::from_slice() reject inputs where this isn't true.
+    // FIXME: Make OpenType::Cmap::from_slice() reject inputs where this isn't true.
     EXPECT_EQ(cmap.glyph_id_for_code_point(0xfeff), 0u);
     EXPECT_EQ(cmap.glyph_id_for_code_point(0xffff), 0xffffu);
     EXPECT_EQ(cmap.glyph_id_for_code_point(0x1'0000), 0u);
