+++ Userland/Libraries/LibGfx/Font/OpenType/Glyf.h
@@ -10,11 +10,11 @@
 #include <AK/Vector.h>
 #include <LibGfx/AffineTransform.h>
 #include <LibGfx/Bitmap.h>
+#include <LibGfx/Font/OpenType/Tables.h>
 #include <LibGfx/Font/PathRasterizer.h>
-#include <LibGfx/Font/TrueType/Tables.h>
 #include <math.h>
 
-namespace TTF {
+namespace OpenType {
 
 class Loca {
 public:
